requirements.txt
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
brazilcep==6.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cssselect==1.2.0
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.112.1
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
isodate==0.7.2
jmespath==1.0.1
loguru==0.7.2
lxml==5.3.0
packaging==24.1
parsel==1.9.1
pi==0.1.2
platformdirs==4.3.6
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
pytest-fastapi==0.1.0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
schematics==2.1.1
sentry-sdk==2.16.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
uvicorn==0.31.1
w3lib==2.2.1
zeep==4.2.1
user-agent==0.1.10
flake8==7.1.1
isort==5.13.2
pylint==3.3.1
tenacity==9.0.0
cachetools==5.5.0
